add nydus image artifact
Artifact manifest is a nice way to offer the ability of enhancing
registry as well as applying minimum modifications to client tools.

Nydus image is an alternative representation of container image,
providing benefits like making starting containers as fast as
possible.

We'd like to have nydus image persist as a new artifact like helm
charts.

Note that we uses loose "reference" for the ociv1 image because our
assumption is that a Nydus image can be stored and used without the
ociv1 image's help.

Signed-off-by: Peng Tao <<tao.peng@linux.alibaba.com>>
Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>