Add nydus image artifact
Artifact manifest is a nice way to offer the ability of enhancing
registry as well as applying minimum modifications to client tools.

Nydus image is an alternative representation of container image,
providing benefits like making starting containers as fast as
possible.

We'd like to have nydus image persist as a new artifact.

Signed-off-by: Liu Bo bo.liu@linux.alibaba.com
Signed-off-by: Peng Tao tao.peng@linux.alibaba.com